Move VerticalLabel from OWConfusion.py to gui.py
