+++ Orange/widgets/gui.py
@@ -3097,6 +3097,38 @@ class ColoredBarItemDelegate(QtGui.QStyledItemDelegate):
             bar_brush = self.color
         return QtGui.QBrush(bar_brush)
 
+
+class VerticalLabel(QtGui.QLabel):
+    def __init__(self, text, parent=None):
+        super().__init__(text, parent)
+        self.setSizePolicy(QtGui.QSizePolicy.Preferred,
+                           QtGui.QSizePolicy.MinimumExpanding)
+        self.setMaximumWidth(self.sizeHint().width() + 2)
+        self.setMargin(4)
+
+    def sizeHint(self):
+        metrics = QtGui.QFontMetrics(self.font())
+        rect = metrics.boundingRect(self.text())
+        size = QtCore.QSize(rect.height() + self.margin(),
+                            rect.width() + self.margin())
+        return size
+
+    def setGeometry(self, rect):
+        super().setGeometry(rect)
+
+    def paintEvent(self, event):
+        painter = QtGui.QPainter(self)
+        rect = self.geometry()
+        text_rect = QtCore.QRect(0, 0, rect.width(), rect.height())
+
+        painter.translate(text_rect.bottomLeft())
+        painter.rotate(-90)
+        painter.drawText(
+            QtCore.QRect(QtCore.QPoint(0, 0),
+                         QtCore.QSize(rect.height(), rect.width())),
+            Qt.AlignCenter, self.text())
+        painter.end()
+
 ##############################################################################
 # progress bar management
